requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain-core
streamlit
langchain-groq
